Move slugify function to lib/util/string
