+++ lib/compose/processors/recompute-codes-voies.js
@@ -1,6 +1,6 @@
 const {first, memoize, chain, minBy} = require('lodash')
 const {createFantoirCommune} = require('@etalab/fantoir')
-const slugify = require('../../util/slugify')
+const {slugify} = require('../../util/string')
 
 const fantoirPath = process.env.FANTOIR_PATH || 'data/fantoir.sqlite'
