+++ lib/util/__tests__/string.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const slugify = require('../slugify')
+const {slugify} = require('../string')
 
 test('slugify()', t => {
   t.is(slugify('Rue du Maréchal Joffre'), 'rue-du-marechal-joffre')
