+++ lib/util/string.js
@@ -4,4 +4,4 @@ function slugify(str) {
   return trim(deburr(str).toLowerCase().replace(/\W+/g, '-'), ' -')
 }
 
-module.exports = slugify
+module.exports = {slugify}
